requirements.txt
sqlalchemy==1.4.22
pyodbc==4.0.30
pyyaml==5.4.1
pytz==2022.1
PyMySQL==1.0.2
tqdm
faker
mysql-connector-python==8.0.27
faker-vehicle==0.2.0
faker_music==0.4
